+++ kwplot/auto_backends.py
@@ -5,7 +5,7 @@ import os
 import ubelt as ub
 
 __all__ = [
-    'autompl', 'autoplt', 'set_mpl_backend', 'BackendContext',
+    'autompl', 'autoplt', 'autosns', 'set_mpl_backend', 'BackendContext',
 ]
 
 
@@ -139,13 +139,14 @@ def autompl(verbose=0, recheck=False, force=None):
         else:
             DISPLAY = os.environ.get('DISPLAY', '')
             if DISPLAY:
-                # Check if we can actually connect to X
-                # NOTE: this call takes a significant amount of time
-                info = ub.cmd('xdpyinfo', shell=True)
-                if verbose:
-                    print('xdpyinfo-info = {}'.format(ub.repr2(info)))
-                if info['ret'] != 0:
-                    DISPLAY = None
+                if sys.platform.startswith('linux') and ub.find_exe('xdpyinfo'):
+                    # On Linux, check if we can actually connect to X
+                    # NOTE: this call takes a significant amount of time
+                    info = ub.cmd('xdpyinfo', shell=True)
+                    if verbose:
+                        print('xdpyinfo-info = {}'.format(ub.repr2(info)))
+                    if info['ret'] != 0:
+                        DISPLAY = None
 
             if verbose:
                 print(' * DISPLAY = {!r}'.format(DISPLAY))
@@ -267,6 +268,17 @@ def autoplt(verbose=0, recheck=False):
     autompl(verbose=verbose, recheck=recheck)
     from matplotlib import pyplot as plt
     return plt
+
+
+def autosns(verbose=0, recheck=False):
+    """
+    Like autosns, but also calls `seaborn.set` and returns the `seaborn` module
+    for convenience.
+    """
+    autompl(verbose=verbose, recheck=recheck)
+    import seaborn as sns
+    sns.set()
+    return sns
 
 
 class BackendContext(object):