+++ CHANGELOG.md
@@ -3,7 +3,12 @@
 This changelog follows the specifications detailed in: [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
-## Version 0.4.8 - Unreleased
+## Version 0.4.9 - Unreleased
+
+### Added
+* `kwplot.autosns` for auto seaborn convenience.
+
+## Version 0.4.8 - Released 2021-04-22
 
 ### Fixed
 * Fixed potential crash in autompl on some systems with conflicting PyQt5 and